requirements.txt
python-telegram-bot
spacy
python-dotenv
wikipedia
openai